2004-03-30	Ralf Corsepius <ralf_corsepius@rtems.org>
	* rtems/include/rtems/rtems/region.h:
	Include stddef.h for size_t.
	(rtems_region_get_segment_size):  Use size_t instead of uint32_t for "size" (3rd arg).
	* rtems/src/regiongetsegmentsize.c: Reflect changes above.
	* libcsupport/src/malloc.c: Reflect changes above.

+++ cpukit/libcsupport/src/malloc.c
@@ -245,7 +245,7 @@ void *malloc(
 #ifdef MALLOC_STATS
   if (return_this)
   {
-      uint32_t   actual_size;
+      size_t     actual_size;
       uint32_t   current_depth;
       status = rtems_region_get_segment_size(
                    RTEMS_Malloc_Heap, return_this, &actual_size);
@@ -369,7 +369,7 @@ void free(
  
 #ifdef MALLOC_STATS
   {
-      uint32_t          size;
+      size_t size;
       status = rtems_region_get_segment_size( RTEMS_Malloc_Heap, ptr, &size );
       if ( status == RTEMS_SUCCESSFUL ) {
           MSBUMP(lifetime_freed, size);
